\name{print.enet}
\alias{print.enet}
\title{Print method for enet objects}
\description{
Print out an enet fit. 
}
\usage{
print.enet(x, ...)  
}
\arguments{
  \item{x}{enet object}
\item{\dots}{Additonal arguments for generic print.}
}
\value{NULL}
\references{
Zou and Hastie (2005) "Regularization and Variable Selection via the Elastic Net"
\emph{Journal of the Royal Statistical Society, Series B,67,301-320}.
}
\author{Hui Zou and Trevor Hastie}
\keyword{methods}
