\name{print.spca}
\alias{print.spca}
\title{Print method for spca objects}
\description{
Print out a spca fit. 
}
\usage{
\method{print}{spca}(x, ...)  
}
\arguments{
  \item{x}{spca object}
\item{\dots}{Additonal arguments for generic print.}
}
\value{NULL}
\references{
Zou, H., Hastie, T. and Tibshirani, R. (2004) "Sparse principal component
analysis" \emph{Technical report, Statistics Dept. Stanford University}.
}
\author{Hui Zou and Trevor Hastie}
\keyword{methods}
