\name{print.arrayspc}
\alias{print.arrayspc}
\title{Print method for arrayspc objects}
\description{
Print out an arrayspc fit. 
}
\usage{
\method{print}{arrayspc}(x, ...)  
}
\arguments{
  \item{x}{arrayspc object}
\item{\dots}{Additonal arguments for generic print.}
}
\value{NULL}
\references{
Zou, H., Hastie, T. and Tibshirani, R. (2004) "Sparse principal component
analysis" \emph{Technical report, Statistics Dept. Stanford University}.
}
\author{Hui Zou and Trevor Hastie}
\keyword{methods}

